fix nivo-slider styling issues
nivo-slider had problems with captions on mobile (they would reach over and outside of the actual image when scaled that small).
Additionally, this commit also changes the caption to use only the `caption` on headers that aren't nivo-slider. This makes the captioning system more consistent with nivo's. If there is a need to use a header on a caption, including an H1 in the caption should suffice.